load gateway servie in gateway wizard
