+++ app/src/main/java/com/kii/gatewaysample/ui/fragments/GatewayServicesListFragment.java
@@ -1,14 +1,10 @@
 package com.kii.gatewaysample.ui.fragments;
 
 import android.app.Activity;
-import android.app.Dialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.DialogFragment;
-import android.support.v7.app.AlertDialog;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -17,15 +13,9 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.kii.gatewaysample.R;
 import com.kii.gatewaysample.model.UPnPService;
-import com.kii.gatewaysample.utils.UPnPControlPointPromise;
-
-import org.jdeferred.DoneCallback;
-import org.jdeferred.FailCallback;
-import org.jdeferred.Promise;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
@@ -42,57 +32,39 @@ public class GatewayServicesListFragment extends DialogFragment {
     ListView listView;
 
     private GatewayServiceArrayAdapter adapter;
-    private UPnPService selectedGateway;
+    private UPnPService selectedService;
+    private UPnPService[] foundServices;
 
     public GatewayServicesListFragment() {
         // Required empty public constructor
     }
 
-    public static GatewayServicesListFragment newFragment(){
+    public static GatewayServicesListFragment newFragment(UPnPService[] foundServices){
         GatewayServicesListFragment fragment = new GatewayServicesListFragment();
+        fragment.foundServices = foundServices;
         return fragment;
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        this.loadGatewayServices();
         View view = inflater.inflate(R.layout.gateway_services_list_fragment, null);
         getDialog().getWindow().requestFeature(Window.FEATURE_NO_TITLE);
         ButterKnife.bind(this, view);
         this.adapter = new GatewayServiceArrayAdapter(getActivity());
         this.listView.setAdapter(adapter);
+        adapter.addAll(this.foundServices);
         this.listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                 String location = ((TextView) view.findViewById(R.id.gateway_ip) ).getText().toString();
-                selectedGateway = (UPnPService)adapter.getItem(position);
+                selectedService = (UPnPService)adapter.getItem(position);
             }
         });
 
         return view;
     }
 
-    private void loadGatewayServices() {
-        new UPnPControlPointPromise(this.getContext()).discover(null).then(new DoneCallback<UPnPService[]>() {
-            @Override
-            public void onDone(UPnPService[] result) {
-                // FIXME: 5/31/16 Can not dismiss well
-                if(result.length == 0){
-                    dismiss();
-                    Toast.makeText(getActivity(), "No gateway found!", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                adapter.clear();
-                adapter.addAll(result);
-            }
-        }).fail(new FailCallback<Throwable>() {
-            @Override
-            public void onFail(Throwable result) {
-                Toast.makeText(getActivity(), result.getMessage(), Toast.LENGTH_SHORT).show();
-            }
-        });
-    }
 
     private class GatewayServiceArrayAdapter extends ArrayAdapter<UPnPService> {
         private final LayoutInflater inflater;
@@ -119,8 +91,8 @@ public class GatewayServicesListFragment extends DialogFragment {
     @OnClick(R.id.select_btn)
     public void onSelect(View v){
         Intent intent = new Intent();
-        if (selectedGateway != null) {
-            intent.putExtra(GATEWAY_SERVICE, selectedGateway);
+        if (selectedService != null) {
+            intent.putExtra(GATEWAY_SERVICE, selectedService);
         }
         getTargetFragment().onActivityResult(getTargetRequestCode(), Activity.RESULT_OK, intent);
         dismiss();
